Added conditions to start game
Cant draw on canvas without starting game by clicking button and choosing the role you want to play.
ALSO switched the variable gameStart to "playerChosen" so I changed all of the instances of the old variable in this code, but make sure to check you don't have any old instances that weren't changed in yours.